Use model.to_param instead of model.id for URLs
Rails uses the #to_param method on model objects in its url helpers, so
identifiers other than the database's primary key can easily be used to
identify a resource in a url. Since, as far as I am aware, JSON API
doesn't have any specific requirements for what an id has to be,
this allows users to use customize their URLs in the same manner as they
would in a Rails view.