Delegate to_json to the serialized hash
Calling `to_json` on a serializer currently produces a different result
depending on which JSON library is providing the method, because of the
convention that `as_json` serializes `self`, rather than its arguments.

Concretely, the problem I ran into is that Active Support's JSON encoder
calls `as_json` but duplicates the provided argument first, which caused
the model I passed to forget that it was persisted:

https://github.com/rails/rails/blob/v5.2.0/activesupport/lib/active_support/json/encoding.rb#L35
https://github.com/rails/rails/blob/v5.2.0/activerecord/lib/active_record/core.rb#L385

By providing our own implementation of `to_json` that delegates to the
result of `as_json`, we can ensure its behaviour is consistent across
different JSON encoders.