Remove kaminari-mongoid dependency
This reverts commit b3117c113776ee81dfa9643f8db5d5242bbaefaa.

Using Kaminari 0.17.0 with Mongoid loaded causes a deprecation warning:

    DEPRECATION WARNING: Kaminari Mongoid support has been extracted to a separate gem, and will be removed in the next 1.0 release. Please bundle kaminari-mongoid gem.

This dependency was added to silence that warning in applications that
use Mongoid, but those applications should add the dependency directly.

Eventually I'd like to relax the Kaminari dependency and allow 1.0 to be
used, but before that we should publish a version that no longer depends
on kaminari-mongoid, so that applications that currently rely on it via
restpack_serializer can see the deprecation warning and bundle the gem.